backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.5
fonttools==4.55.1
h11==0.14.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pydantic==2.10.2
pydantic_core==2.27.1
pyFirmata==1.1.0
pyparsing==3.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
